+++ components/Modal/WritePost/WritePost.tsx
@@ -1,10 +1,10 @@
 import useModalStore from '@stores/useModalStore';
 
-import ImageUploader from '@components/ImageUploader/ImageUploader';
+import ImageUploader from '@components/ImageUploader';
 
 import styles from './WritePost.module.scss';
 import useCreateForm from '../../../app/map/_hooks/useCreateForm';
-import ModalContainer from '../Modal';
+import Modal from '../Modal';
 
 export default function WritePost() {
   const { toggleModal } = useModalStore();
@@ -12,23 +12,23 @@ export default function WritePost() {
   // TODO: 카테고리는 구체화 될때까지 보류, 위치도 보류
 
   return (
-    <ModalContainer>
+    <Modal>
       <form onSubmit={handleSubmit(onSubmit)}>
-        <ModalContainer.Header>글 작성</ModalContainer.Header>
-        <ModalContainer.Body>
+        <Modal.Header>글 작성</Modal.Header>
+        <Modal.Body>
           <label htmlFor='title'>title</label>
           <input id='title' {...register('title')} type='text' placeholder='제목' className={styles.input} />
           <label htmlFor='content'>content</label>
           <input id='content' {...register('content')} type='text' placeholder='내용' className={styles.input} />
           <ImageUploader setValue={setValue} />
-        </ModalContainer.Body>
-        <ModalContainer.Footer>
+        </Modal.Body>
+        <Modal.Footer>
           <button type='button' onClick={cancelForm} className={styles.button}>
             취소
           </button>
           <button type='submit'>완료</button>
-        </ModalContainer.Footer>
+        </Modal.Footer>
       </form>
-    </ModalContainer>
+    </Modal>
   );
 }
